daily-language-be/requirements.txt
django
django-cors-headers
djangorestframework
gunicorn
python-decouple